Working customizable svg envelopes
