+++ packages/audio-components/src/elements/controls/EnvelopeSVG.ts
@@ -1,14 +1,23 @@
-// EnvelopeSVG.ts - Clean embeddable component
+// EnvelopeSVG.ts
 import van from '@repo/vanjs-core';
-import { AudioEnvelopeController } from './CustomEnvElement'; // Import just the controller
+import { type CustomEnvelope, DEFAULT_ENV } from '@repo/audiolib';
 
-const { svg, path, line, g } = van.tags('http://www.w3.org/2000/svg');
+const { svg, path, line, g, div } = van.tags('http://www.w3.org/2000/svg');
 
 export const EnvelopeSVG = (
-  envelopeController: AudioEnvelopeController,
+  envelope: CustomEnvelope | null,
   width: string = '100%',
   height: string = '120px'
 ) => {
+  if (!envelope) {
+    return div(
+      {
+        style: `width: ${width}; height: ${height}; background: #1a1a1a; border: 1px solid #444; border-radius: 4px; display: flex; align-items: center; justify-content: center; color: #666;`,
+      },
+      'Loading envelope...'
+    );
+  }
+
   // UI states
   const selectedPoint = van.state<number | null>(null);
   const isDragging = van.state(false);
@@ -16,11 +25,18 @@ export const EnvelopeSVG = (
   let svgElement: SVGSVGElement;
   let pointsGroup: SVGGElement;
 
+  // UI update trigger
+  const updateTrigger = van.state(0);
+
+  const unsubscribe = envelope?.onChange(() => {
+    updateTrigger.val++;
+  });
+
   const updateControlPoints = () => {
     if (!pointsGroup) return;
 
     pointsGroup.innerHTML = '';
-    const points = envelopeController.getPoints();
+    const points = envelope?.getPoints() || DEFAULT_ENV;
 
     points.forEach((point, index) => {
       const circle = document.createElementNS(
@@ -47,7 +63,7 @@ export const EnvelopeSVG = (
 
       if (index > 0 && index < points.length - 1) {
         circle.addEventListener('dblclick', () => {
-          envelopeController.deletePoint(index);
+          envelope?.deletePoint(index);
         });
       }
 
@@ -66,7 +82,7 @@ export const EnvelopeSVG = (
         0,
         Math.min(1, 1 - (e.clientY - rect.top) / rect.height)
       );
-      envelopeController.updatePoint(selectedPoint.val, time, value);
+      envelope?.updatePoint(selectedPoint.val, time, value);
     }
   };
 
@@ -83,7 +99,7 @@ export const EnvelopeSVG = (
       0,
       Math.min(1, 1 - (e.clientY - rect.top) / rect.height)
     );
-    envelopeController.addPoint(time, value);
+    envelope?.addPoint(time, value);
   };
 
   // Create SVG
@@ -128,8 +144,8 @@ export const EnvelopeSVG = (
   // Envelope path
   const envelopePath = path({
     d: () => {
-      envelopeController.updateTrigger.val;
-      return envelopeController.getSVGPath(400, 200);
+      updateTrigger.val;
+      return envelope?.getSVGPath(400, 200);
     },
     fill: 'none',
     stroke: '#4ade80',
@@ -147,10 +163,160 @@ export const EnvelopeSVG = (
 
   // Reactivity
   van.derive(() => {
-    envelopeController.updateTrigger.val;
+    updateTrigger.val;
     selectedPoint.val;
     updateControlPoints();
   });
 
   return svgElement;
 };
+
+// IGNORE BELOW
+// export const EnvelopeSVG = (
+//   envelope: CustomEnvelope | null,
+//   width: string = '100%',
+//   height: string = '120px'
+// ) => {
+//   if (!envelope) {
+//     return div(
+//       {
+//         style: `width: ${width}; height: ${height}; background: #1a1a1a; border: 1px solid #444; border-radius: 4px; display: flex; align-items: center; justify-content: center; color: #666;`,
+//       },
+//       'Loading envelope...'
+//     );
+//   }
+
+//   const selectedPoint = van.state<number | null>(null);
+//   const isDragging = van.state(false);
+//   const updateTrigger = van.state(0);
+
+//   const unsubscribe = envelope.onChange(() => {
+//     updateTrigger.val++;
+//   });
+
+//   const { circle } = van.tags('http://www.w3.org/2000/svg');
+
+//   // Create a container that we can dynamically update
+//   const controlPointsContainer = g({ class: 'control-points' });
+
+//   // Function to update control points manually
+//   const updateControlPoints = () => {
+//     // Clear existing points
+//     controlPointsContainer.innerHTML = '';
+
+//     const points = envelope.getPoints() || DEFAULT_ENV;
+//     points.forEach((point, index) => {
+//       const circleElement = circle({
+//         cx: point.time * 400,
+//         cy: (1 - point.value) * 200,
+//         r: 4,
+//         fill: selectedPoint.val === index ? '#ff6b6b' : '#4ade80',
+//         stroke: '#fff',
+//         'stroke-width': 1,
+//         style: 'cursor: pointer;',
+//         onmousedown: (e: MouseEvent) => {
+//           selectedPoint.val = index;
+//           isDragging.val = true;
+//           e.preventDefault();
+//         },
+//         ...(index > 0 &&
+//           index < points.length - 1 && {
+//             ondblclick: () => envelope.deletePoint(index),
+//           }),
+//       });
+
+//       van.add(controlPointsContainer, circleElement);
+//     });
+//   };
+
+//   // Set up reactivity using van.derive
+//   van.derive(() => {
+//     updateTrigger.val; // Dependency
+//     selectedPoint.val; // Dependency
+//     updateControlPoints();
+//   });
+
+//   return svg(
+//     {
+//       viewBox: '0 0 400 200',
+//       preserveAspectRatio: 'none',
+//       style: `width: ${width}; height: ${height}; background: #1a1a1a; border: 1px solid #444; border-radius: 4px;`,
+//       onmousemove: (e: MouseEvent) => {
+//         if (isDragging.val && selectedPoint.val !== null) {
+//           const rect = (e.currentTarget as SVGElement).getBoundingClientRect();
+//           const time = Math.max(
+//             0,
+//             Math.min(1, (e.clientX - rect.left) / rect.width)
+//           );
+//           const value = Math.max(
+//             0,
+//             Math.min(1, 1 - (e.clientY - rect.top) / rect.height)
+//           );
+//           envelope.updatePoint(selectedPoint.val, time, value);
+//         }
+//       },
+//       onmouseup: () => {
+//         isDragging.val = false;
+//         selectedPoint.val = null;
+//       },
+//       onmouseleave: () => {
+//         isDragging.val = false;
+//         selectedPoint.val = null;
+//       },
+//       ondblclick: (e: MouseEvent) => {
+//         if (isDragging.val) return;
+//         const rect = (e.currentTarget as SVGElement).getBoundingClientRect();
+//         const time = Math.max(
+//           0,
+//           Math.min(1, (e.clientX - rect.left) / rect.width)
+//         );
+//         const value = Math.max(
+//           0,
+//           Math.min(1, 1 - (e.clientY - rect.top) / rect.height)
+//         );
+//         envelope.addPoint(time, value);
+//       },
+//     },
+
+//     // Grid
+//     g(
+//       { class: 'grid' },
+//       ...Array.from({ length: 6 }, (_, i) => {
+//         const x = (i / 5) * 400;
+//         return line({
+//           x1: x,
+//           y1: 0,
+//           x2: x,
+//           y2: 200,
+//           stroke: '#333',
+//           'stroke-width': 1,
+//         });
+//       }),
+//       ...Array.from({ length: 6 }, (_, i) => {
+//         const y = (i / 5) * 200;
+//         return line({
+//           x1: 0,
+//           y1: y,
+//           x2: 400,
+//           y2: y,
+//           stroke: '#333',
+//           'stroke-width': 1,
+//         });
+//       })
+//     ),
+
+//     // Envelope path
+//     path({
+//       d: () => {
+//         updateTrigger.val;
+//         return envelope.getSVGPath(400, 200);
+//       },
+//       fill: 'none',
+//       stroke: '#4ade80',
+//       'stroke-width': 2,
+//     }),
+
+//     // Control points container
+//     controlPointsContainer
+//   );
+// };
