+++ packages/audio-components/src/index.ts
@@ -3,4 +3,4 @@ export type { Audiolib, AudiolibOptions } from '@repo/audiolib';
 
 export { defineKarplusSynth } from './elements/instruments/KarplusSynthElement';
 export { defineSampler } from './elements/instruments/SamplerElement';
-export { defineEnvelope } from './elements/controls/CustomEnvElement';
+export { defineEnvelope } from './elements/controls/EnvelopeElement';
