+++ packages/audio-components/src/utils/vanjs-utils.ts
@@ -1,0 +1,8 @@
+// todo: maybe add this to vanjs util
+export const when = (condition: any, elementFn: () => any) =>
+  condition ? elementFn() : null;
+
+// export const when = <T>(
+//   condition: T,
+//   elementFn: (value: NonNullable<T>) => any
+// ) => (condition ? elementFn(condition as NonNullable<T>) : null);
