+++ packages/audiolib/src/nodes/params/index.ts
@@ -1,3 +1,4 @@
+export * from './Envelope';
 export * from './macros/MacroParam';
 export * from './macros/AudioParamController';
 export * from './helpers/ValueSnapper';
